Added Wiper.lua
Script for FPV camera lens servo wiper.

On Taranis, go to Custom Scripts and add this LUA. Set duration and delay. If you want different settings e.g. on a switch, assign this lua as many times as needed. After that, go to Mixer page and put this/these as input(s).